+++ src/client/dmo/mod.rs
@@ -1,0 +1,4 @@
+pub use self::{operation::DmoOperation, response::DmoResponse};
+
+mod operation;
+mod response;